Add better parsing for RFC test vectors
