+++ Cargo.toml
@@ -51,11 +51,13 @@ bincode = "1"
 chacha20poly1305 = "0.8"
 criterion = "0.3"
 hex = "0.4"
+json = "0.12"
 lazy_static = "1"
 opaque-ke = { path = "", default-features = false, features = ["std"] }
 serde_json = "1"
 sha2 = "0.9"
 proptest = "1"
+regex = "1"
 rustyline = "8"
 
 [[bench]]
