+++ src/tests/mod.rs
@@ -5,5 +5,7 @@
 
 mod full_test;
 pub mod mock_rng;
-mod opaque_test_vectors;
+mod opaque_vectors;
+mod parser;
+mod test_opaque_vectors;
 mod voprf_test_vectors;
